Use core as external dep, move domain classes out
