+++ app/src/main/java/org/coepi/android/cross/ScannedTcnsHandler.kt
@@ -3,10 +3,10 @@ package org.coepi.android.cross
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.rxkotlin.plusAssign
 import io.reactivex.rxkotlin.subscribeBy
-import org.coepi.android.core.ObservedTcnsRecorder
 import org.coepi.android.ble.BleManager
 import org.coepi.android.system.log.log
 import org.coepi.android.ui.debug.DebugBleObservable
+import org.coepi.core.services.ObservedTcnsRecorder
 
 class ScannedTcnsHandler(
     bleManager: BleManager,
