+++ app/src/main/java/org/coepi/android/domain/symptomflow/SymptomFlow.kt
@@ -1,14 +1,14 @@
 package org.coepi.android.domain.symptomflow
 
-import org.coepi.android.domain.symptomflow.SymptomId.BREATHLESSNESS
-import org.coepi.android.domain.symptomflow.SymptomId.COUGH
-import org.coepi.android.domain.symptomflow.SymptomId.DIARRHEA
-import org.coepi.android.domain.symptomflow.SymptomId.FEVER
-import org.coepi.android.domain.symptomflow.SymptomId.LOSS_SMELL_OR_TASTE
-import org.coepi.android.domain.symptomflow.SymptomId.MUSCLE_ACHES
-import org.coepi.android.domain.symptomflow.SymptomId.NONE
-import org.coepi.android.domain.symptomflow.SymptomId.OTHER
-import org.coepi.android.domain.symptomflow.SymptomId.RUNNY_NOSE
+import org.coepi.core.domain.model.SymptomId.BREATHLESSNESS
+import org.coepi.core.domain.model.SymptomId.COUGH
+import org.coepi.core.domain.model.SymptomId.DIARRHEA
+import org.coepi.core.domain.model.SymptomId.FEVER
+import org.coepi.core.domain.model.SymptomId.LOSS_SMELL_OR_TASTE
+import org.coepi.core.domain.model.SymptomId.MUSCLE_ACHES
+import org.coepi.core.domain.model.SymptomId.NONE
+import org.coepi.core.domain.model.SymptomId.OTHER
+import org.coepi.core.domain.model.SymptomId.RUNNY_NOSE
 import org.coepi.android.domain.symptomflow.SymptomStep.BREATHLESSNESS_DESCRIPTION
 import org.coepi.android.domain.symptomflow.SymptomStep.COUGH_DAYS
 import org.coepi.android.domain.symptomflow.SymptomStep.COUGH_DESCRIPTION
@@ -20,6 +20,7 @@ import org.coepi.android.domain.symptomflow.SymptomStep.FEVER_TEMPERATURE_SPOT
 import org.coepi.android.domain.symptomflow.SymptomStep.FEVER_TEMPERATURE_TAKEN_TODAY
 import org.coepi.android.extensions.add
 import org.coepi.android.system.log.log
+import org.coepi.core.domain.model.SymptomId
 
 class SymptomFlow(private var steps: List<SymptomStep>) {
     init {
