+++ app/src/main/java/org/coepi/android/extensions/DateExtensions.kt
@@ -1,6 +1,6 @@
 package org.coepi.android.extensions
 
-import org.coepi.android.domain.UnixTime
+import org.coepi.core.domain.model.UnixTime
 import java.util.Date
 
 fun Date.toUnixTime() = UnixTime.fromDate(this)
