+++ app/src/main/java/org/coepi/android/extensions/InstantExtensions.kt
@@ -1,6 +1,6 @@
 package org.coepi.android.extensions
 
-import org.coepi.android.domain.UnixTime
+import org.coepi.core.domain.model.UnixTime
 import org.threeten.bp.Instant
 
 fun Instant.toUnixTime() = UnixTime.fromInstant(this)
