+++ app/src/main/java/org/coepi/android/repo/SymptomRepo.kt
@@ -12,15 +12,15 @@ import org.coepi.android.R.string.symptom_report_title_no_symptoms
 import org.coepi.android.R.string.symptom_report_title_other
 import org.coepi.android.R.string.symptom_report_title_runny_nose
 import org.coepi.android.domain.model.Symptom
-import org.coepi.android.domain.symptomflow.SymptomId.BREATHLESSNESS
-import org.coepi.android.domain.symptomflow.SymptomId.COUGH
-import org.coepi.android.domain.symptomflow.SymptomId.DIARRHEA
-import org.coepi.android.domain.symptomflow.SymptomId.FEVER
-import org.coepi.android.domain.symptomflow.SymptomId.LOSS_SMELL_OR_TASTE
-import org.coepi.android.domain.symptomflow.SymptomId.MUSCLE_ACHES
-import org.coepi.android.domain.symptomflow.SymptomId.NONE
-import org.coepi.android.domain.symptomflow.SymptomId.OTHER
-import org.coepi.android.domain.symptomflow.SymptomId.RUNNY_NOSE
+import org.coepi.core.domain.model.SymptomId.BREATHLESSNESS
+import org.coepi.core.domain.model.SymptomId.COUGH
+import org.coepi.core.domain.model.SymptomId.DIARRHEA
+import org.coepi.core.domain.model.SymptomId.FEVER
+import org.coepi.core.domain.model.SymptomId.LOSS_SMELL_OR_TASTE
+import org.coepi.core.domain.model.SymptomId.MUSCLE_ACHES
+import org.coepi.core.domain.model.SymptomId.NONE
+import org.coepi.core.domain.model.SymptomId.OTHER
+import org.coepi.core.domain.model.SymptomId.RUNNY_NOSE
 import org.coepi.android.system.Resources
 
 interface SymptomRepo {
