+++ app/src/main/java/org/coepi/android/tcn/AlertsDao.kt
@@ -1,6 +1,7 @@
 package org.coepi.android.tcn
 
 import io.reactivex.Observable
+import org.coepi.core.domain.model.Alert
 
 interface AlertsDao {
     val alerts: Observable<List<Alert>>
