+++ app/src/main/java/org/coepi/android/tcn/RealmAlertDao.kt
@@ -2,12 +2,17 @@ package org.coepi.android.tcn
 
 import io.reactivex.subjects.BehaviorSubject
 import io.realm.RealmResults
-import io.realm.kotlin.createObject
 import io.realm.kotlin.where
-import org.coepi.android.domain.UnixTime
-import org.coepi.android.domain.symptomflow.UserInput
 import org.coepi.android.repo.RealmProvider
 import org.coepi.android.system.log.log
+import org.coepi.core.domain.model.Alert
+import org.coepi.core.domain.model.toInt
+import io.realm.kotlin.createObject
+import org.coepi.core.domain.model.UnixTime
+import org.coepi.core.domain.model.UserInput.None
+import org.coepi.core.domain.model.UserInput.Some
+import org.coepi.core.domain.model.toCoughSeverity
+import org.coepi.core.domain.model.toFeverSeverity
 
 class RealmAlertDao(private val realmProvider: RealmProvider) : AlertsDao {
     private val realm get() = realmProvider.realm
@@ -39,9 +44,9 @@ class RealmAlertDao(private val realmProvider: RealmProvider) : AlertsDao {
 
             realmObj.contactTime = alert.contactTime.value
             realmObj.reportTime = alert.reportTime.value
-            realmObj.earliestSymptomTime = when (alert.earliestSymptomTime) {
-                is UserInput.Some -> alert.earliestSymptomTime.value.value
-                is UserInput.None -> null
+            realmObj.earliestSymptomTime = when (val earliestSymptomTime = alert.earliestSymptomTime) {
+                is Some -> earliestSymptomTime.value.value
+                is None -> null
             }
             realmObj.feverSeverity = alert.feverSeverity.toInt()
             realmObj.coughSeverity = alert.coughSeverity.toInt()
@@ -87,8 +92,8 @@ class RealmAlertDao(private val realmProvider: RealmProvider) : AlertsDao {
         Alert(
             id = id,
             reportTime = UnixTime.fromValue(reportTime),
-            earliestSymptomTime = earliestSymptomTime?.let { UserInput.Some(UnixTime.fromValue(it)) }
-                ?: UserInput.None,
+            earliestSymptomTime = earliestSymptomTime?.let { Some(UnixTime.fromValue(it)) }
+                ?: None,
             feverSeverity = toFeverSeverity(feverSeverity),
             coughSeverity = toCoughSeverity(coughSeverity),
             breathlessness = breathlessness,
@@ -99,6 +104,5 @@ class RealmAlertDao(private val realmProvider: RealmProvider) : AlertsDao {
             other = other,
             noSymptoms = noSymptoms,
             contactTime = UnixTime.fromValue(contactTime)
-
         )
 }
