+++ app/src/main/java/org/coepi/android/ui/symptoms/breathless/BreathlessViewData.kt
@@ -1,6 +1,7 @@
 package org.coepi.android.ui.symptoms.breathless
 
 import android.graphics.drawable.Drawable
-import org.coepi.android.domain.symptomflow.SymptomInputs.Breathlessness
+import org.coepi.core.domain.model.SymptomInputs.Breathlessness
 
-data class BreathlessViewData(val name: String, val icon: Drawable?, val isChecked: Boolean, val breathless: Breathlessness.Cause)
+data class BreathlessViewData(val name: String, val icon: Drawable?, val isChecked: Boolean,
+                              val breathless: Breathlessness.Cause)
