+++ app/src/main/java/org/coepi/android/ui/symptoms/cough/CoughStatusViewData.kt
@@ -1,5 +1,5 @@
 package org.coepi.android.ui.symptoms.cough
 
-import org.coepi.android.domain.symptomflow.SymptomInputs.Cough
+import org.coepi.core.domain.model.SymptomInputs.Cough
 
 data class CoughStatusViewData(val name: String, val isChecked: Boolean, val status: Cough.Status)
