+++ app/src/main/java/org/coepi/android/ui/symptoms/cough/CoughTypeViewModel.kt
@@ -2,19 +2,17 @@ package org.coepi.android.ui.symptoms.cough
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.android.schedulers.AndroidSchedulers.mainThread
 import org.coepi.android.domain.symptomflow.SymptomFlowManager
-import org.coepi.android.domain.symptomflow.SymptomInputs.Cough.Type.DRY
-import org.coepi.android.domain.symptomflow.SymptomInputs.Cough.Type.WET
-import org.coepi.android.domain.symptomflow.UserInput
-import org.coepi.android.domain.symptomflow.UserInput.Some
 import org.coepi.android.extensions.rx.toIsInProgress
 import org.coepi.android.extensions.rx.toLiveData
 import org.coepi.android.ui.navigation.NavigationCommand.Back
 import org.coepi.android.ui.navigation.RootNavigation
+import org.coepi.core.domain.model.SymptomInputs.Cough.Type.DRY
+import org.coepi.core.domain.model.SymptomInputs.Cough.Type.WET
+import org.coepi.core.domain.model.UserInput.Some
 
-class CoughTypeViewModel (
+class CoughTypeViewModel(
     private val navigation: RootNavigation,
     private val symptomFlowManager: SymptomFlowManager
 ) : ViewModel() {
