+++ app/src/main/java/org/coepi/android/ui/symptoms/fever/FeverHighestTemperatureViewModel.kt
@@ -2,16 +2,15 @@ package org.coepi.android.ui.symptoms.fever
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.android.schedulers.AndroidSchedulers.mainThread
-import org.coepi.android.domain.model.Temperature.Fahrenheit
 import org.coepi.android.domain.symptomflow.SymptomFlowManager
-import org.coepi.android.domain.symptomflow.UserInput.None
-import org.coepi.android.domain.symptomflow.UserInput.Some
 import org.coepi.android.extensions.rx.toIsInProgress
 import org.coepi.android.extensions.rx.toLiveData
 import org.coepi.android.ui.navigation.NavigationCommand.Back
 import org.coepi.android.ui.navigation.RootNavigation
+import org.coepi.core.domain.model.Temperature.Fahrenheit
+import org.coepi.core.domain.model.UserInput.None
+import org.coepi.core.domain.model.UserInput.Some
 
 class FeverHighestTemperatureViewModel(
     val navigation: RootNavigation,
