+++ app/src/main/java/org/coepi/android/ui/symptoms/fever/FeverTakenTodayViewModel.kt
@@ -5,11 +5,11 @@ import androidx.lifecycle.ViewModel
 import io.reactivex.android.schedulers.AndroidSchedulers.mainThread
 import org.coepi.android.domain.symptomflow.SymptomFlowManager
 import org.coepi.android.domain.symptomflow.SymptomStep.FEVER_HIGHEST_TEMPERATURE
-import org.coepi.android.domain.symptomflow.UserInput.Some
 import org.coepi.android.extensions.rx.toIsInProgress
 import org.coepi.android.extensions.rx.toLiveData
 import org.coepi.android.ui.navigation.NavigationCommand.Back
 import org.coepi.android.ui.navigation.RootNavigation
+import org.coepi.core.domain.model.UserInput.Some
 
 class FeverTakenTodayViewModel(
     private val navigation: RootNavigation,
