+++ app/src/main/java/org/coepi/android/ui/symptoms/fever/FeverTemperatureSpotViewModel.kt
@@ -2,19 +2,18 @@ package org.coepi.android.ui.symptoms.fever
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
-import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.android.schedulers.AndroidSchedulers.mainThread
 import org.coepi.android.NavGraphRootDirections.Companion.actionGlobalFeverTemperatureSpotInputFragment
 import org.coepi.android.domain.symptomflow.SymptomFlowManager
-import org.coepi.android.domain.symptomflow.SymptomInputs.Fever.TemperatureSpot.Armpit
-import org.coepi.android.domain.symptomflow.SymptomInputs.Fever.TemperatureSpot.Ear
-import org.coepi.android.domain.symptomflow.SymptomInputs.Fever.TemperatureSpot.Mouth
-import org.coepi.android.domain.symptomflow.UserInput.Some
 import org.coepi.android.extensions.rx.toIsInProgress
 import org.coepi.android.extensions.rx.toLiveData
 import org.coepi.android.ui.navigation.NavigationCommand.Back
 import org.coepi.android.ui.navigation.NavigationCommand.ToDestination
 import org.coepi.android.ui.navigation.RootNavigation
+import org.coepi.core.domain.model.SymptomInputs.Fever.TemperatureSpot.Armpit
+import org.coepi.core.domain.model.SymptomInputs.Fever.TemperatureSpot.Ear
+import org.coepi.core.domain.model.SymptomInputs.Fever.TemperatureSpot.Mouth
+import org.coepi.core.domain.model.UserInput.Some
 
 class FeverTemperatureSpotViewModel (
     private val navigation: RootNavigation,
